EnterpriseVoiceBot/requirements.txt
Flask
boto3>=1.26.159
langchain
python-dotenv
urllib